LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/media-elements/loading-the-media-resource/resource-selection-invoke-load-expected.txt


FAIL invoking resource selection with load() assert_equals: networkState after v.load() expected 3 but got 0
